pyproject.toml
[tool.poetry]
name = "healthcheck-threshold-eval"
version = "0.1.0"
description = ""
authors = ["minseong"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.2.0"
openpyxl = "^3.1.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
tqdm = "^4.0.0"
rich = "^13.0.0"
jsonlines = "^4.0.0"
openai = "^1.0.0"
numpy = "^2.4.3"
scikit-learn = "^1.8.0"
sentence-transformers = "^5.3.0"
transformers = "^5.3.0"
torch = "^2.10.0"
accelerate = "^1.13.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
pytest = "^8.0.0"
ipykernel = "^6.0.0"
ruff = "^0.15.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"